+++ models/gpt2/transformer.py
@@ -1,30 +1,17 @@
-from dataclasses import dataclass
 import math
 
 import torch
 from torch import nn, Tensor
 import torch.nn.functional as F
-from torch.nn import Embedding, Linear, LayerNorm
-
-@dataclass
-class GPTConfig:
-  vocab_size: int = 50257
-  n_ctx: int = 1024
-  n_layer: int = 12
-  n_head: int = 12
-  n_embd: int = 768
-  norm_eps: float = 1e-5
 
 
 def _attention(q: Tensor, k: Tensor, v: Tensor, mask: Tensor, scale: float) -> Tensor:
-  mask = mask.to(q.device)
   att = (q @ k.transpose(-2, -1)) * scale # (B, H, T, T)
   att = att.masked_fill(mask < 0.5, -float('inf')) # causal mask
   y = F.softmax(att, dim=-1) @ v # (B, H, T, T) x (B, H, T, C') -> (B, H, T, C')
   return y
 
 def _fused_attention(q: Tensor, k: Tensor, v: Tensor, mask: Tensor, scale: float) -> Tensor:
-  mask = mask.to(q.device)
   y = F.scaled_dot_product_attention(q, k, v, mask > 0.5, scale=scale) # flash attention
   return y
 
@@ -34,18 +21,19 @@ class CausalSelfAttention(nn.Module):
     super().__init__()
     assert n_embd % n_head == 0
     self.n_head, self.n_embd, self.head_size = n_head, n_embd, n_embd // n_head
-    self.c_attn = Linear(n_embd, 3 * n_embd)
-    self.c_proj = Linear(n_embd, n_embd)
+    self.c_attn = nn.Linear(n_embd, 3 * n_embd)
+    self.c_proj = nn.Linear(n_embd, n_embd)
     self.bias = torch.tril(torch.ones(1, 1, n_ctx, n_ctx))
 
   def forward(self, x: Tensor):
     B, T, C= x.size()
     qkv = self.c_attn(x) # (B, T, 3C)
-    q, k, v = qkv.split(self.n_embd, dim=2) # (B, T, C)
+    q, k, v = qkv.split(self.n_embd, dim=-1) # (B, T, C)
     q = q.view(B, T, self.n_head, self.head_size).transpose(1, 2) # (B, H, T, C')
     k = k.view(B, T, self.n_head, self.head_size).transpose(1, 2) # (B, H, T, C')
     v = v.view(B, T, self.n_head, self.head_size).transpose(1, 2) # (B, H, T, C')
-    y = _fused_attention(q, k, v, self.bias[:,:,:T,:T], (1.0 / math.sqrt(self.head_size))) # (B, H, T, C')
+    mask = self.bias[:,:,:T,:T].to(q.device)
+    y = _fused_attention(q, k, v, mask, (1.0 / math.sqrt(self.head_size))) # (B, H, T, C')
     y = y.transpose(1, 2).contiguous().view(B, T, C)
     y = self.c_proj(y)
     return y
@@ -54,8 +42,8 @@ class CausalSelfAttention(nn.Module):
 class MLP(nn.Module):
   def __init__(self, n_embd: int):
     super().__init__()
-    self.c_fc = Linear(n_embd, n_embd * 4)
-    self.c_proj = Linear(4 * n_embd, n_embd)
+    self.c_fc = nn.Linear(n_embd, n_embd * 4)
+    self.c_proj = nn.Linear(4 * n_embd, n_embd)
 
   def forward(self, x: Tensor):
     x = self.c_proj(F.gelu(self.c_fc(x), approximate='tanh'))
@@ -65,9 +53,9 @@ class MLP(nn.Module):
 class Block(nn.Module):
   def __init__(self, n_embd: int, n_head: int, n_ctx: int, norm_eps: float):
     super().__init__()
-    self.ln_1 = LayerNorm(n_embd, eps=norm_eps)
+    self.ln_1 = nn.LayerNorm(n_embd, eps=norm_eps)
     self.attn = CausalSelfAttention(n_embd, n_head, n_ctx)
-    self.ln_2 = LayerNorm(n_embd, eps=norm_eps)
+    self.ln_2 = nn.LayerNorm(n_embd, eps=norm_eps)
     self.mlp = MLP(n_embd)
 
   def forward(self, x: Tensor):
@@ -80,10 +68,10 @@ class Transformer(nn.Module):
   def __init__(self, n_embd: int, n_head: int, n_ctx: int, norm_eps: float, vocab_size: int, n_layer: int, **_):
     super().__init__()
     self.transformer = nn.ModuleDict(dict(
-      wte = Embedding(vocab_size, n_embd),
-      wpe = Embedding(n_ctx, n_embd),
+      wte = nn.Embedding(vocab_size, n_embd),
+      wpe = nn.Embedding(n_ctx, n_embd),
       h = nn.ModuleList([Block(n_embd, n_head, n_ctx, norm_eps) for _ in range(n_layer)]),
-      ln_f = LayerNorm(n_embd, eps=norm_eps),
+      ln_f = nn.LayerNorm(n_embd, eps=norm_eps),
     ))
     self.lm_head = nn.Linear(n_embd, vocab_size, bias=False)
     self.apply_weight_sharing()
