+++ models/llama2/generate.py
@@ -5,7 +5,7 @@ from tqdm import tqdm
 import torch
 import torch.nn.functional as F
 
-from models.helpers import timeit
+from models.helpers import Generator, timeit
 from models.llama.tokenizer import Tokenizer
 from models.llama2.transformer import Transformer
 from models.llama.config import LlamaConfig
@@ -13,16 +13,9 @@ from models.llama2.load import build
 from models.llama.generate import sample_top_p
 
 
-class Llama:
+class Llama(Generator):
   def __init__(self, model: Transformer, tokenizer: Tokenizer, config: LlamaConfig):
     self.model, self.tokenizer, self.config = model, tokenizer, config
-
-  @property
-  def device(self) -> torch.device: return next(self.model.parameters()).device
-
-  def to(self, device: torch.device):
-    self.model = self.model.to(device)
-    return self
 
   @staticmethod
   @timeit(desc="Load time", ms=False)
@@ -176,53 +169,3 @@ def text_completion(generator: Llama, prompts: List[str], temperature: float = 0
       for t, logprobs_i in zip(generation_tokens, generation_logprobs)
     ]
   return [{"generation": tokenizer.decode(t)} for t in generation_tokens]
-
-
-@torch.inference_mode
-def text_completion2(generator: Llama, prompts: List[str], temperature: float=0.6,
-                    top_p: float=0.9, max_gen_len: Optional[int]=None):
-  model, tokenizer = generator.model, generator.tokenizer,
-  max_batch_size, max_seq_len = generator.config.max_batch_size, generator.config.max_seq_len
-  device = generator.device
-  if max_gen_len is None:
-    max_gen_len = max_seq_len - 1
-  prompt_tokens = [tokenizer.encode(prompt, bos=True, eos=False) for prompt in prompts]
-  bsz = len(prompt_tokens)
-  assert bsz <= max_batch_size, f"batch size must be less than or equal to {max_batch_size}"
-  min_prompt_len = min(len(prompt) for prompt in prompt_tokens)
-  max_prompt_len = max(len(prompt) for prompt in prompt_tokens)
-  assert max_prompt_len <= max_seq_len, f"prompt length must be less than or equal to {max_seq_len}"
-  total_len = min(max_seq_len, max_gen_len + max_prompt_len)
-  pad_id = tokenizer.pad_id
-  tokens = torch.full((bsz, total_len), pad_id, dtype=torch.long, device=device)
-  prev_pos = 0
-  for k, t in enumerate(prompt_tokens):
-    tokens[k, : len(t)] = torch.tensor(t, dtype=torch.long, device=device)
-  eos_reached = torch.tensor([False] * bsz, device=device)
-  tokens_mask = tokens != pad_id # True if the token is a prompt token, False otherwise
-  model.eval()
-  for cur_pos in tqdm(range(min_prompt_len, total_len), desc='Generating tokens'):
-    with torch.no_grad():
-      logits = model(tokens[:, prev_pos:cur_pos], prev_pos)
-    if temperature > 0:
-      probs = torch.softmax(logits[:, -1] / temperature, dim=-1)
-      next_token = sample_top_p(probs, top_p)
-    else:
-      next_token = torch.argmax(logits[:, -1], dim=-1) # greedy
-    prev_pos = cur_pos
-    next_token = next_token.reshape(-1)
-    # Only replace token if it is a padding token
-    next_token = torch.where(tokens_mask[:, cur_pos], tokens[:, cur_pos], next_token)
-    tokens[:, cur_pos] = next_token
-    # EOS is reached only if we found an EOS token for a padding position
-    eos_reached |= (~tokens_mask[:, cur_pos]) & (next_token == tokenizer.eos_id)
-    if all(eos_reached): break
-  out_tokens, out_text = [], []
-  for current_prompt_tokens in tokens.tolist():
-    # Cut to the EOS token, if present
-    if tokenizer.eos_id in current_prompt_tokens:
-      eos_idx = current_prompt_tokens.index(tokenizer.eos_id)
-      current_prompt_tokens = current_prompt_tokens[:eos_idx]
-    out_tokens.append(current_prompt_tokens)
-    out_text.append(tokenizer.decode(current_prompt_tokens))
-  return out_tokens, out_text
